GEOMESA-2630,GEOMESA-2631 Fix JSON output for Spark dataframes
* New conversion class in spark-sql module
* Deprecated spark-jts conversion class
* Separated out classes in GeoMesaSparkSQL file
