+++ docs/user/spark/core.rst
@@ -139,3 +139,20 @@ To save features, use the ``save()`` method:
 Note that some providers may be read-only.
 
 See :doc:`./providers` for details on specific provider implementations.
+
+
+GeoJSON Output
+^^^^^^^^^^^^^^
+
+The ``geomesa-spark-core`` module provides a means of exporting an ``RDD[SimpleFeature]`` to a
+`GeoJSON <http://geojson.org/>`__ string. This allows for quick visualization of the data in many front-end mapping
+libraries that support GeoJSON input such as Leaflet or Open Layers.
+
+To convert an RDD, import the implicit conversion and invoke the ``asGeoJSONString`` method.
+
+.. code-block:: scala
+
+    import org.locationtech.geomesa.spark.SpatialRDD._
+    val rdd: RDD[SimpleFeature] = ???
+    val geojson = rdd.asGeoJSONString
+
